# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Sample exchnage API",
      "description": "A sample exchange api",
      "version": "1.0"
    },
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "securityDefinitions": {
      "Bearer": {
          "type": "apiKey",
          "name": "Authorization",
          "in": "header"
      }
    },
    "definitions": {
      "Pet": {
        "required": [
          "login",
          "password",
          "enable",
          "balance",
          "email"
        ],
        "properties": {
          "login": {
            "type": "string",
            "example": "doggie"
          },
          "password": {
            "type": "string",
            "example": "asd"
          },
          "enable": {
            "type": "bool",
            "example": true
          },
          "balance": {
            "type": "float",
            "example": "1000"
          },
          "email": {
            "type": "string",
            "example": "name@gmail.com"
          }
        }
      }
    },
    "paths": {
      "/.version": {
        "get": {
          "security": [
            {"Bearer": []}
          ],
          "summary": "get verions.",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Successful response.",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/oauth2/token": {
        "get": {
          "summary": "Generate bearer token.",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Successful response.",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/symbols/": {
        "get": {
          "security": [
            {"Bearer": []}
          ],
          "summary": "Get symbol value.",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Successful response.",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/quotes/{symbol}": {
        "get": {
          "security": [
            {"Bearer": []}
          ],
          "summary": "Get symbol value.",
          "produces": [
            "application/json"
          ],
          "parameters":[
            {
              "name":"symbol",
              "in":"path",
              "description":"The symbol that needs to be fetched",
              "required":true,
              "type":"string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response.",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/accounts/": {
        "get": {
          "security": [
            {"Bearer": []}
          ],
          "summary": "Get account info.",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Successful response.",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "post": {
          "security": [
            {"Bearer": []}
          ],
          "summary": "Get account info.",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Pet object that needs to be added to the store",
              "required": false,
              "schema": {
                "$ref": "#/definitions/Pet"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response.",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
}